test added

+++ src/main/java/com/aimcodes/fashionBlog/services/serviceImpl/CommentServiceImpl.java
@@ -42,31 +42,31 @@ public class CommentServiceImpl implements CommentService {
             String uuidGen = uuidGenerator.generateUuid();
             Comment comment = Comment.builder()
                     .content(request.getContent())
-                            .user(user)
-                                    .post(post)
-                                        .comment_author(user.getUsername())
-                                            .uuid(uuidGen)
-                                                .build();
+                    .user(user)
+                    .post(post)
+                    .comment_author(user.getUsername())
+                    .uuid(uuidGen)
+                    .build();
             commentRepository.save(comment);
 
             CommentResponseDto response = CommentResponseDto.builder()
                     .content(comment.getContent())
-                        .author(comment.getComment_author())
-                            .build();
+                    .author(comment.getComment_author())
+                    .build();
 
             return new ResponseEntity<>(responseManager.successfulRequest(response), HttpStatus.ACCEPTED);
 
         }else if(user == null && request.getComment_author() != null){
             Comment comment = Comment.builder()
                     .content(request.getContent())
-                            .comment_author(request.getComment_author())
-                                .post(post).build();
+                    .comment_author(request.getComment_author())
+                    .post(post).build();
             commentRepository.save(comment);
 
             CommentResponseDto response = CommentResponseDto.builder()
                     .content(comment.getContent())
-                        .author(comment.getComment_author())
-                            .build();
+                    .author(comment.getComment_author())
+                    .build();
 
             return new ResponseEntity<>(responseManager.successfulRequest(response), HttpStatus.ACCEPTED);
         }
@@ -80,8 +80,8 @@ public class CommentServiceImpl implements CommentService {
 
         CommentResponseDto response = CommentResponseDto.builder()
                 .content(comment.getContent())
-                    .author(comment.getComment_author())
-                        .build();
+                .author(comment.getComment_author())
+                .build();
         if(post == null)
             throw new HandleNullException("Invalid post", "Post with id " + uuid + " does not exist in the database");
         return new ResponseEntity<>(responseManager.successfulRequest(response), HttpStatus.ACCEPTED);
@@ -92,14 +92,13 @@ public class CommentServiceImpl implements CommentService {
         User user = (User) session.getAttribute("currUser");
         Comment comment = commentRepository.findByUuid(uuid);
 
-        if(user != null && user.getRole().equals(Role.valueOf("ADMIN"))){
+        if(user != null && user.getRole().equals(Role.valueOf("ADMIN")) && comment != null){
             commentRepository.delete(comment);
-
             return new ResponseEntity<>(responseManager.successfulRequest("comment deleted successfully!"), HttpStatus.ACCEPTED);
         } else if (user == null)
             throw new HandleNullException("Invalid user", "No user in session");
-        if(comment == null)
-            throw new HandleNullException("Comment unavailable", "Comment with id " + uuid + " does not exist in the database");
+        else if(comment == null)
+             throw new HandleNullException("Comment unavailable", "Comment with id " + uuid + " does not exist in the database");
         throw new NoAccessException("No access to delete comment", "wrong user or user is not an admin");
     }
 
@@ -111,9 +110,9 @@ public class CommentServiceImpl implements CommentService {
         if(comment != null) {
             CommentResponseDto response = CommentResponseDto.builder()
                     .author(comment.getComment_author())
-                        .content(comment.getContent())
-                            .likes(comment.getLikes().size())
-                                .build();
+                    .content(comment.getContent())
+                    .likes(comment.getLikes().size())
+                    .build();
             return new ResponseEntity<>(responseManager.successfulRequest(response), HttpStatus.FOUND);
         }
         throw new HandleNullException("Comment unavailable", "Comment with id " + uuid + " does not exist in the database");
@@ -130,9 +129,9 @@ public class CommentServiceImpl implements CommentService {
             comments.forEach(comment -> {
                 CommentResponseDto response = CommentResponseDto.builder()
                         .author(comment.getComment_author())
-                                .content(comment.getContent())
-                                        .likes(comment.getLikes().size())
-                                                .build();
+                        .content(comment.getContent())
+                        .likes(comment.getLikes().size())
+                        .build();
                 responses.add(response);
 
             });
